backend/requirements.txt
fastapi[all]
sqlmodel
psycopg2-binary
python-jose[cryptography]
types-python-jose
passlib[bcrypt]
bcrypt==4.0.1